+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.3.3 / 5.58.3] - 2022-08-??
 
+### Changed
+- improved sandman settings behavioure for non admin users [#2123](https://github.com/sandboxie-plus/Sandboxie/issues/2123)
+
 ### Fixed
 - fixed issues with group moving drag and drop
 - approved more requied sys calls
@@ -14,6 +17,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - fixed issue with config protection [#2206](https://github.com/sandboxie-plus/Sandboxie/issues/2206)
 - fixed issue with default box [#2195](https://github.com/sandboxie-plus/Sandboxie/issues/2195)
 - fixed issue with keyboard delete shortcut for process termination
+
+
 
 
 ## [1.3.2 / 5.58.2] - 2022-08-30